requirements.txt
django~=5.0
djangorestframework~=3.14.0
celery~=5.3.0
requests~=2.31.0
pyyaml~=6.0.0
django-rest-passwordreset>=1.3.0
